:art: auditId字段改为Long

+++ weixin-java-miniapp/src/main/java/cn/binarywang/wx/miniapp/bean/live/WxMaLiveResult.java
@@ -20,7 +20,7 @@ import java.util.List;
 public class WxMaLiveResult implements Serializable {
   private static final long serialVersionUID = 1L;
   private Integer total;
-  private Integer auditId;
+  private Long auditId;
   private Integer goodsId;
   private List<Goods> goods;
